+++ Chapter2/src/index.js
@@ -20,6 +20,20 @@ app.post("/accounts", (request, response) => {
     });
   }
 
+  app.get("/statements/:cpf", (request, response) => {
+    const { cpf } = request.params;
+
+    const customer = customers.find((customer) => customer.cpf === cpf);
+
+    if (!customer) {
+      return response.status(404).json({
+        error: "Customer not found.",
+      });
+    }
+
+    return response.json(customer.statement);
+  });
+
   const createdCustomer = { id: uuidv4(), name, cpf, statement: [] };
 
   customers.push(createdCustomer);